Feat: #64 메타데이터 삭제, webp 변환 + internalApplication 삭제
