+++ convert/src/main/java/image/module/convert/DataClient.java
@@ -1,8 +1,7 @@
-package image.module.internal;
+package image.module.convert;
 
-import image.module.internal.dto.ImageRequest;
+import image.module.convert.dto.ImageRequest;
 import org.springframework.cloud.openfeign.FeignClient;
-import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
