+++ convert/src/main/java/image/module/convert/config/MinioConfig.java
@@ -1,7 +1,6 @@
-package image.module.internal.config;
+package image.module.convert.config;
 
 import io.minio.MinioClient;
-import lombok.Getter;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
