+++ convert/src/main/java/image/module/convert/dto/ImageRequest.java
@@ -1,14 +1,11 @@
-package image.module.internal.dto;
+package image.module.convert.dto;
 
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 
-import javax.imageio.ImageIO;
-import java.awt.image.BufferedImage;
 import java.io.File;
-import java.io.IOException;
 import java.util.UUID;
 
 @Getter
