+++ convert/src/main/resources/application-dev.yml
@@ -1,12 +1,12 @@
 spring:
   application:
     name: convert
-#  kafka:
-#    bootstrap-servers: kafka:29092
-#    consumer:
-#      group-id: image-upload-group # 컨슈머 그룹 ID
-#      key-deserializer: org.apache.kafka.common.serialization.StringDeserializer  # 키 역직렬화
-#      value-deserializer: org.apache.kafka.common.serialization.StringDeserializer  # 값 역직렬화
+  kafka:
+    bootstrap-servers: kafka:29092
+    consumer:
+      group-id: image-upload-group # 컨슈머 그룹 ID
+      key-deserializer: org.apache.kafka.common.serialization.StringDeserializer  # 키 역직렬화
+      value-deserializer: org.apache.kafka.common.serialization.StringDeserializer  # 값 역직렬화
 server:
   port: 19097
