--- internal/build.gradle
@@ -1,70 +1,0 @@
-plugins {
-	id 'java'
-	id 'org.springframework.boot' version '3.3.4'
-	id 'io.spring.dependency-management' version '1.1.6'
-}
-
-group = 'image.module'
-version = '0.0.1-SNAPSHOT'
-
-java {
-	toolchain {
-		languageVersion = JavaLanguageVersion.of(17)
-	}
-}
-
-configurations {
-	compileOnly {
-		extendsFrom annotationProcessor
-	}
-}
-
-repositories {
-	mavenCentral()
-}
-
-ext {
-	set('springCloudVersion', "2023.0.3")
-}
-
-dependencies {
-	// feignClient
-	implementation 'org.springframework.cloud:spring-cloud-starter-openfeign'
-
-	// kafka
-	implementation 'org.springframework.kafka:spring-kafka'
-
-	// EurekaClient
-	implementation 'org.springframework.cloud:spring-cloud-starter-netflix-eureka-client'
-
-	// Thumbnailator (이미지 리사이징)
-	implementation group: 'net.coobird', name: 'thumbnailator', version: '0.4.20'
-
-	//minio
-	implementation 'io.minio:minio:8.5.0'
-	implementation 'com.squareup.okhttp3:okhttp:4.9.3'
-
-	// Image Processing (WebP)
-	implementation "com.sksamuel.scrimage:scrimage-core:4.0.32"
-	implementation "com.sksamuel.scrimage:scrimage-webp:4.0.32"
-
-	// Apache Commons IO / ImageService에서 FileUtils 사용 시 필요 (파일 복사)
-	implementation 'commons-io:commons-io:2.11.0'
-
-	implementation 'org.springframework.boot:spring-boot-starter-web'
-	compileOnly 'org.projectlombok:lombok'
-	annotationProcessor 'org.projectlombok:lombok'
-	testImplementation 'org.springframework.boot:spring-boot-starter-test'
-	testImplementation 'org.springframework.kafka:spring-kafka-test'
-	testRuntimeOnly 'org.junit.platform:junit-platform-launcher'
-}
-
-dependencyManagement {
-	imports {
-		mavenBom "org.springframework.cloud:spring-cloud-dependencies:${springCloudVersion}"
-	}
-}
-
-tasks.named('test') {
-	useJUnitPlatform()
-}
