--- internal/src/main/java/image/module/internal/InternalApplication.java
@@ -1,15 +1,0 @@
-package image.module.internal;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-import org.springframework.cloud.openfeign.EnableFeignClients;
-
-@EnableFeignClients
-@SpringBootApplication
-public class InternalApplication {
-
-	public static void main(String[] args) {
-		SpringApplication.run(InternalApplication.class, args);
-	}
-
-}
