--- internal/src/main/resources/application-prod.yml
@@ -1,35 +1,0 @@
-spring:
-  application:
-    name: internal
-  kafka:
-    bootstrap-servers: ${KAFKA_URL}
-    producer:
-      key-serializer: org.apache.kafka.common.serialization.StringSerializer
-      value-serializer: org.apache.kafka.common.serialization.StringSerializer
-server:
-  port: 19094
-
-eureka:
-  client:
-    service-url:
-      defaultZone: ${EUREKA_URL}
-
-logging:
-  level:
-    org:
-      hibernate:
-        orm:
-          jdbc:
-            bind: trace
-
-minio:
-  url: ${MINIO_URL}
-  access:
-    key: minio
-    secret: minioadmin
-  buckets:
-    downloadBucket: mybucket
-    uploadBucket: update-bucket
-
-cdn-server:
-  url: ${CDN_URL}
