--- internal/src/test/java/image/module/internal/InternalApplicationTests.java
@@ -1,13 +1,0 @@
-package image.module.internal;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class InternalApplicationTests {
-
-	@Test
-	void contextLoads() {
-	}
-
-}
